4.15-4.23.

+++ controllers/blogs.js
@@ -1,17 +1,34 @@
 const blogsRouter = require('express').Router()
+const config = require('../utils/config')
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const logger = require('../utils/logger')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1})
   response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
-  //console.log(`new blog to be added: ${blog}`)
-  if (!blog.likes) blog.likes = 0
+  const body = request.body
+  const user = request.user
+
+  if (!user) {
+    return response.status(401).json({ error: 'userId missing or not valid' })
+  }
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: !body.likes ? 0 : body.likes,
+    user: user._id
+  })
 
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
   response.status(201).json(savedBlog)
 })
 
@@ -32,8 +49,21 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const user = request.user
+
+  if (!user) {
+    return response.status(400).json({ error: 'userId missing or not valid' })
+  }
+
+  const deletableBlog = await Blog.findById(request.params.id)
+  const blogUserID = deletableBlog.user._id.toString()
+
+  if (user.id === blogUserID) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({ error: 'user invalid' })
+  }
 })
 
 module.exports = blogsRouter
